+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Select input implementations
 - suffix and prefix text input enable
 - radio input implementation
+- check box and check group components
 
 ## 2021.11.16